module: move helpers out of cjs loader
PR-URL: https://github.com/nodejs/node/pull/49912
