+++ lib/internal/modules/cjs/loader.js
@@ -55,7 +55,6 @@ const {
   StringPrototypeCharAt,
   StringPrototypeCharCodeAt,
   StringPrototypeEndsWith,
-  StringPrototypeLastIndexOf,
   StringPrototypeIndexOf,
   StringPrototypeRepeat,
   StringPrototypeSlice,
@@ -68,7 +67,7 @@ const cjsParseCache = new SafeWeakMap();
 
 // Set first due to cycle with ESM loader functions.
 module.exports = {
-  wrapSafe, Module, toRealPath, readPackageScope, cjsParseCache,
+  wrapSafe, Module, cjsParseCache,
   get hasLoadedAnyUserCJSModule() { return hasLoadedAnyUserCJSModule; },
   initializeCJS,
 };
@@ -88,9 +87,7 @@ const {
 const { internalCompileFunction } = require('internal/vm');
 const assert = require('internal/assert');
 const fs = require('fs');
-const internalFS = require('internal/fs/utils');
 const path = require('path');
-const { sep } = path;
 const { internalModuleStat } = internalBinding('fs');
 const { safeGetenv } = internalBinding('credentials');
 const {
@@ -106,6 +103,7 @@ const {
   makeRequireFunction,
   normalizeReferrerURL,
   stripBOM,
+  toRealPath,
 } = require('internal/modules/helpers');
 const packageJsonReader = require('internal/modules/package_json_reader');
 const { getOptionValue, getEmbedderOptions } = require('internal/options');
@@ -403,15 +401,7 @@ function initializeCJS() {
 //   -> a.<ext>
 //   -> a/index.<ext>
 
-/**
- * @param {string} requestPath
- * @return {PackageConfig}
- */
-function readPackage(requestPath) {
-  return packageJsonReader.read(path.resolve(requestPath, 'package.json'));
-}
-
-let _readPackage = readPackage;
+let _readPackage = packageJsonReader.readPackage;
 ObjectDefineProperty(Module, '_readPackage', {
   __proto__: null,
   get() { return _readPackage; },
@@ -422,37 +412,6 @@ ObjectDefineProperty(Module, '_readPackage', {
   },
   configurable: true,
 });
-
-/**
- * Get the nearest parent package.json file from a given path.
- * Return the package.json data and the path to the package.json file, or false.
- * @param {string} checkPath The path to start searching from.
- */
-function readPackageScope(checkPath) {
-  const rootSeparatorIndex = StringPrototypeIndexOf(checkPath, sep);
-  let separatorIndex;
-  const enabledPermission = permission.isEnabled();
-  do {
-    separatorIndex = StringPrototypeLastIndexOf(checkPath, sep);
-    checkPath = StringPrototypeSlice(checkPath, 0, separatorIndex);
-    // Stop the search when the process doesn't have permissions
-    // to walk upwards
-    if (enabledPermission && !permission.has('fs.read', checkPath + sep)) {
-      return false;
-    }
-    if (StringPrototypeEndsWith(checkPath, sep + 'node_modules')) {
-      return false;
-    }
-    const pjson = _readPackage(checkPath + sep);
-    if (pjson.exists) {
-      return {
-        data: pjson,
-        path: checkPath,
-      };
-    }
-  } while (separatorIndex > rootSeparatorIndex);
-  return false;
-}
 
 /**
  * Try to load a specifier as a package.
@@ -499,14 +458,6 @@ function tryPackage(requestPath, exts, isMain, originalPath) {
 }
 
 /**
- * Cache for storing resolved real paths of modules.
- * In order to minimize unnecessary lstat() calls, this cache is a list of known-real paths.
- * Set to an empty Map to reset.
- * @type {Map<string, string>}
- */
-const realpathCache = new SafeMap();
-
-/**
  * Check if the file exists and is not a directory if using `--preserve-symlinks` and `isMain` is false, keep symlinks
  * intact, otherwise resolve to the absolute realpath.
  * @param {string} requestPath The path to the file to load.
@@ -519,17 +470,6 @@ function tryFile(requestPath, isMain) {
     return path.resolve(requestPath);
   }
   return toRealPath(requestPath);
-}
-
-
-/**
- * Resolves the path of a given `require` specifier, following symlinks.
- * @param {string} requestPath The `require` specifier
- */
-function toRealPath(requestPath) {
-  return fs.realpathSync(requestPath, {
-    [internalFS.realpathCacheKey]: realpathCache,
-  });
 }
 
 /**
@@ -593,7 +533,7 @@ function trySelfParentPath(parent) {
 function trySelf(parentPath, request) {
   if (!parentPath) { return false; }
 
-  const { data: pkg, path: pkgPath } = readPackageScope(parentPath);
+  const { data: pkg, path: pkgPath } = packageJsonReader.readPackageScope(parentPath);
   if (!pkg || pkg.exports == null || pkg.name === undefined) {
     return false;
   }
@@ -1153,7 +1093,7 @@ Module._resolveFilename = function(request, parent, isMain, options) {
 
   if (request[0] === '#' && (parent?.filename || parent?.id === '<repl>')) {
     const parentPath = parent?.filename ?? process.cwd() + path.sep;
-    const pkg = readPackageScope(parentPath) || { __proto__: null };
+    const pkg = packageJsonReader.readPackageScope(parentPath) || { __proto__: null };
     if (pkg.data?.imports != null) {
       try {
         const { packageImportsResolve } = require('internal/modules/esm/resolve');
@@ -1450,7 +1390,7 @@ Module._extensions['.js'] = function(module, filename) {
     content = fs.readFileSync(filename, 'utf8');
   }
   if (StringPrototypeEndsWith(filename, '.js')) {
-    const pkg = readPackageScope(filename) || { __proto__: null };
+    const pkg = packageJsonReader.readPackageScope(filename) || { __proto__: null };
     // Function require shouldn't be used in ES modules.
     if (pkg.data?.type === 'module') {
       const parent = moduleParentCache.get(module);
