+++ lib/internal/modules/helpers.js
@@ -21,6 +21,8 @@ const {
 const { BuiltinModule } = require('internal/bootstrap/realm');
 
 const { validateString } = require('internal/validators');
+const fs = require('fs'); // Import all of `fs` so that it can be monkey-patched.
+const internalFS = require('internal/fs/utils');
 const path = require('path');
 const { pathToFileURL, fileURLToPath, URL } = require('internal/url');
 
@@ -38,6 +40,23 @@ let debug = require('internal/util/debuglog').debuglog('module', (fn) => {
 });
 
 /** @typedef {import('internal/modules/cjs/loader.js').Module} Module */
+
+/**
+ * Cache for storing resolved real paths of modules.
+ * In order to minimize unnecessary lstat() calls, this cache is a list of known-real paths.
+ * Set to an empty Map to reset.
+ * @type {Map<string, string>}
+ */
+const realpathCache = new SafeMap();
+/**
+ * Resolves the path of a given `require` specifier, following symlinks.
+ * @param {string} requestPath The `require` specifier
+ */
+function toRealPath(requestPath) {
+  return fs.realpathSync(requestPath, {
+    [internalFS.realpathCacheKey]: realpathCache,
+  });
+}
 
 /** @type {Set<string>} */
 let cjsConditions;
@@ -310,4 +329,5 @@ module.exports = {
   makeRequireFunction,
   normalizeReferrerURL,
   stripBOM,
+  toRealPath,
 };
