+++ lib/internal/modules/run_main.js
@@ -15,12 +15,13 @@ function resolveMainPath(main) {
   // Note extension resolution for the main entry point can be deprecated in a
   // future major.
   // Module._findPath is monkey-patchable here.
-  const { Module, toRealPath } = require('internal/modules/cjs/loader');
+  const { Module } = require('internal/modules/cjs/loader');
   let mainPath = Module._findPath(path.resolve(main), null, true);
   if (!mainPath) { return; }
 
   const preserveSymlinksMain = getOptionValue('--preserve-symlinks-main');
   if (!preserveSymlinksMain) {
+    const { toRealPath } = require('internal/modules/helpers');
     mainPath = toRealPath(mainPath);
   }
 
@@ -48,7 +49,7 @@ function shouldUseESMLoader(mainPath) {
   if (mainPath && StringPrototypeEndsWith(mainPath, '.mjs')) { return true; }
   if (!mainPath || StringPrototypeEndsWith(mainPath, '.cjs')) { return false; }
 
-  const { readPackageScope } = require('internal/modules/cjs/loader');
+  const { readPackageScope } = require('internal/modules/package_json_reader');
   const pkg = readPackageScope(mainPath);
   // No need to guard `pkg` as it can only be an object or `false`.
   return pkg.data?.type === 'module' || getOptionValue('--experimental-default-type') === 'module';
